+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 
 ### New features
 
+- Add `\BSP\DrWatson\DrWatson`
+- Add `\BSP\DrWatson\ExceptionType`
+
 ### Bugfixes
 
 </details>